// src/api/models/CreateAppLocationSettingRequest.ts
/**
 * B2B_API
 * API to access Odin Fleet, empowering developers to deploy gameservers all around the world in just a few lines of code.
 *
 * OpenAPI spec version: 0.8.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Placement } from '../models/Placement.ts';
import { HttpFile } from '../http/http.ts';

export class CreateAppLocationSettingRequest {
    /**
    * The name of the location setting
    */
    'name': string;
    /**
    * The id of the server configuration that should be deployed
    */
    'serverConfigId': number;
    /**
    * Whether auto scaler is enabled
    */
    'autoScalerEnabled': boolean;
    /**
    * The minimum number of instances that should run at the specific location
    */
    'autoScalerMin'?: number | null;
    /**
    * The maximum number of instances that should run at the specific location
    */
    'autoScalerMax'?: number | null;
    /**
    * The number of instances that should run at the specific location
    */
    'numInstances': number;
    /**
    * The placement settings that define the location and other constraints
    */
    'placement'?: Placement;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "serverConfigId",
            "baseName": "serverConfigId",
            "type": "number",
            "format": ""
        },
        {
            "name": "autoScalerEnabled",
            "baseName": "autoScalerEnabled",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "autoScalerMin",
            "baseName": "autoScalerMin",
            "type": "number",
            "format": ""
        },
        {
            "name": "autoScalerMax",
            "baseName": "autoScalerMax",
            "type": "number",
            "format": ""
        },
        {
            "name": "numInstances",
            "baseName": "numInstances",
            "type": "number",
            "format": ""
        },
        {
            "name": "placement",
            "baseName": "placement",
            "type": "Placement",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateAppLocationSettingRequest.attributeTypeMap;
    }

    public constructor() {
    }
}